Add oqtane client services
