+++ Oqtane.Client/Extensions/OqtaneServiceCollectionExtensions.cs
@@ -1,4 +1,13 @@
-﻿using Microsoft.AspNetCore.Components.Authorization;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.IO.Compression;
+using System.Linq;
+using System.Net.Http;
+using System.Reflection;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Components.Authorization;
+using Oqtane.Modules;
 using Oqtane.Providers;
 using Oqtane.Services;
 using Oqtane.Shared;
@@ -45,6 +54,95 @@ namespace Microsoft.Extensions.DependencyInjection
             services.AddScoped<ISystemService, SystemService>();
 
             return services;
+        }
+
+        public static IServiceCollection AddOqtaneClientServices(this IServiceCollection services, HttpClient httpClient)
+        {
+            LoadClientAssemblies(httpClient).GetAwaiter().GetResult();
+
+            // dynamically register module contexts and repository services
+            Assembly[] assemblies = AppDomain.CurrentDomain.GetAssemblies();
+            foreach (Assembly assembly in assemblies)
+            {
+                var implementationTypes = assembly.GetTypes()
+                    .Where(item => item.GetInterfaces().Contains(typeof(IService)));
+
+                foreach (Type implementationtype in implementationTypes)
+                {
+                    Type servicetype = Type.GetType(implementationtype.AssemblyQualifiedName.Replace(implementationtype.Name, "I" + implementationtype.Name));
+                    if (servicetype != null)
+                    {
+                        services.AddScoped(servicetype, implementationtype); // traditional service interface
+                    }
+                    else
+                    {
+                        services.AddScoped(implementationtype, implementationtype); // no interface defined for service
+                    }
+                }
+
+                assembly.GetInstances<IClientStartup>()
+                    .ToList()
+                    .ForEach(x => x.ConfigureServices(services));
+            }
+
+            return services;
+        }
+
+        private static async Task LoadClientAssemblies(HttpClient http)
+        {
+            // get list of loaded assemblies on the client 
+            var assemblies = AppDomain.CurrentDomain.GetAssemblies().Select(a => a.GetName().Name).ToList();
+
+            // get assemblies from server and load into client app domain
+            var zip = await http.GetByteArrayAsync($"/~/api/Installation/load");
+
+            // asemblies and debug symbols are packaged in a zip file
+            using (ZipArchive archive = new ZipArchive(new MemoryStream(zip)))
+            {
+                var dlls = new Dictionary<string, byte[]>();
+                var pdbs = new Dictionary<string, byte[]>();
+
+                foreach (ZipArchiveEntry entry in archive.Entries)
+                {
+                    if (!assemblies.Contains(Path.GetFileNameWithoutExtension(entry.Name)))
+                    {
+                        using (var memoryStream = new MemoryStream())
+                        {
+                            entry.Open().CopyTo(memoryStream);
+                            byte[] file = memoryStream.ToArray();
+                            switch (Path.GetExtension(entry.Name))
+                            {
+                                case ".dll":
+                                    // Loads the stallite assemblies early
+                                    if (entry.Name.EndsWith(Constants.StalliteAssemblyExtension))
+                                    {
+                                        Assembly.Load(file);
+                                    }
+                                    else
+                                    {
+                                        dlls.Add(entry.Name, file);
+                                    }
+                                    break;
+                                case ".pdb":
+                                    pdbs.Add(entry.Name, file);
+                                    break;
+                            }
+                        }
+                    }
+                }
+
+                foreach (var item in dlls)
+                {
+                    if (pdbs.ContainsKey(item.Key))
+                    {
+                        Assembly.Load(item.Value, pdbs[item.Key]);
+                    }
+                    else
+                    {
+                        Assembly.Load(item.Value);
+                    }
+                }
+            }
         }
     }
 }
